Split the implementation of supercall
