+++ kernel/patch/patch.c
@@ -48,7 +48,6 @@ int hotpatch_init();
 int bypass_kcfi();
 int bypass_selinux();
 int resolve_pt_regs();
-int supercall_install();
 void module_init();
 void syscall_init();
 int kstorage_init();
@@ -78,9 +77,6 @@ static void before_rest_init(hook_fargs4_t *args, void *udata)
 
     if ((rc = task_observer())) goto out;
     log_boot("task_observer done: %d\n", rc);
-
-    rc = supercall_install();
-    log_boot("supercall_install done: %d\n", rc);
 
     rc = kstorage_init();
     log_boot("kstorage_init done: %d\n", rc);
