+++ kernel/supercall/Makefile
@@ -1,0 +1,30 @@
+ifndef TARGET_COMPILE
+    $(error TARGET_COMPILE not set)
+endif
+
+ifndef KP_DIR
+    KP_DIR = ../..
+endif
+
+
+CC = $(TARGET_COMPILE)gcc
+LD = $(TARGET_COMPILE)ld
+
+INCLUDE_DIRS := . include patch/include linux/include linux/arch/arm64/include linux/tools/arch/arm64/include
+
+INCLUDE_FLAGS := $(foreach dir,$(INCLUDE_DIRS),-I$(KP_DIR)/kernel/$(dir))
+
+objs := supercall.o
+
+all: supercall.kpm
+
+supercall.kpm: ${objs}
+	${CC} -r -o $@ $^
+
+%.o: %.c
+	${CC} $(CFLAGS) $(INCLUDE_FLAGS) -c -O2 -o $@ $<
+
+.PHONY: clean
+clean:
+	rm -rf *.kpm
+	find . -name "*.o" | xargs rm -f
